requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-axes==6.0.0
psycopg2-binary==2.9.7
redis==5.0.0
django-redis==5.3.0
argon2-cffi==23.1.0
cryptography==42.0.0
django-environ==0.11.2
requests==2.31.0
